Added AJAX Call
Site now uses AJAX to call an API for a random food. Currently, it returns a food title and image irrespective of the user's responses in the questionnaire. Proper functionality will be added later.

Old code has been left in but commented out. It can be used as a reference or as a fall back in case the API falls through.